% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpw.R
\docType{data}
\name{Riverside}
\alias{Riverside}
\title{NEWWS Riverside data}
\value{
A list containing
\item{trunc_dep12sm2}{Outcome. Maternal depression among participants at the end of two years after treatment.}
\item{emp}{Mediator. A binary indicator for whether one was employed in any quarter during the 2 years after randomization.}
\item{emp_cat}{Mediator. A categorical variable for one's employment during the 2 years after randomization.}
\item{emp_cont}{Mediator. A continuous variable for the number of quarters one was employed during the 2 years after randomization.}
\item{treat}{Treatment}
\item{emp_prior}{Mediator}
\item{pqtrunc25}{Preference for taking care of family full time rather than working}
\item{pqtrunc30}{Too many family problems for full-time or part-time job}
\item{pqtrunc49}{Cannot go school training due to too much to do}
\item{pqtrunc50}{Sad past week}
\item{pqtrunc51}{Depressed past week}
\item{pqtrunc52}{Blues past week}
\item{pqtrunc53}{Lonely past week}
\item{hispanic}{Hispanic = 1; otherwise = 0}
\item{nevmar}{Never married = 1; otherwise = 0}
\item{nohsdip}{No high school diploma or GED = 1; otherwise = 0}
\item{AFDC3660}{On AFDC 36 of past 60 months = 1; otherwise = 0}
\item{AFDC0_Y1}{Ever in a situation of not receiving welfare during the first year after randomization = 1; otherwise = 0}
}
\description{
The National Evaluation of Welfare-to-Work Strategies (NEWWS) Riverside study. 
Immediately preceding the welfare reform
nationwide in the mid-1990s, welfare applicants in Riverside, California, were
assigned at random to either a labor force attachment (LFA) program (Z = 1) or
a control condition (Z = 0) in an experimental study. The LFA program, with the
goal of eventually weaning participants from the welfare system, emphasized
seeking and securing employment, offered job search services, and provided
incentives including a threat of sanctions should one fail to meet the program
requirements, while the control group members were guaranteed cash assistance
without the requirement for employment.
}
