% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpw.R
\name{rmpw}
\alias{rmpw}
\title{Causal Mediation Analysis Using Weighting Approach}
\usage{
rmpw(
  data,
  treatment,
  mediator,
  outcome,
  propensity_x,
  outcome_x,
  m.scale,
  t.rand = TRUE
)
}
\arguments{
\item{data}{The data set for analysis.}

\item{treatment}{The name of the treatment variable (string).}

\item{mediator}{The name of the mediator variable (string).}

\item{outcome}{The name of the outcome variable (string).}

\item{propensity_x}{A vector of variable names (string) of pretreatment confounders, which will be included in the propensity score model. To reduce the risk of model misspecifications, we account for the interaction between the treatment and each observed pretreatment covariate in the propensity score model.}

\item{outcome_x}{A vector of variable names (string) of pretreatment confounders, which will be included in the outcome model.}

\item{m.scale}{Scale of the mediator ("discrete" or "continuous").}

\item{t.rand}{A logical value. If TRUE, treatment is randomized. If FALSE, treatment is not randomized. The default is TRUE.}
}
\value{
A list contains the estimation and inference results of the causal effects (including Natural Direct Effect, Natural Indirect Effect, Pure Indirect Effect, Total Direct Effect, and Natural Treatment-by-Mediator Interaction Effect) and the coefficients of the pretreatment covariates. When the mediator is binary, and the treatment is randomized, the ratio-of-mediator-probability weight (Hong et al., 2015) is used with closed-form standard errors that account for the estimation of the weights (Bein et al., 2018). Otherwise, the weight is constructed in the algebraically equivalent inverse-probability form of Huber (2014), which accommodates continuous and multi-valued discrete mediators, and standard errors and confidence intervals are obtained via bootstrapping.
}
\description{
Causal Mediation Analysis Using Weighting Approach
}
\examples{
\donttest{
data(Riverside)
rmpw(data = Riverside, treatment = "treat", mediator = "emp",
    outcome = "trunc_dep12sm2", propensity_x = c("emp_prior",
        "pqtrunc50", "pqtrunc51", "pqtrunc52", "pqtrunc53", "pqtrunc30",
        "hispanic", "pqtrunc49", "nevmar"), outcome_x = c("emp_prior",
        "pqtrunc50", "pqtrunc51", "pqtrunc52", "pqtrunc53", "pqtrunc30",
        "hispanic", "pqtrunc49", "nevmar"), m.scale = "discrete")
rmpw(data = Riverside, treatment = "treat", mediator = "emp_cont",
    outcome = "trunc_dep12sm2", propensity_x = c("emp_prior",
        "pqtrunc50", "pqtrunc51", "pqtrunc52", "pqtrunc53", "pqtrunc30",
        "hispanic", "pqtrunc49", "nevmar"), outcome_x = c("emp_prior",
        "pqtrunc50", "pqtrunc51", "pqtrunc52", "pqtrunc53", "pqtrunc30",
        "hispanic", "pqtrunc49", "nevmar"), m.scale = "continuous")
}
}
\references{
Hong, G., Deutsch, J., & Hill, H. D. (2015). Ratio-of-mediator-probability weighting for causal mediation analysis in the presence of treatment-by-mediator interaction. Journal of Educational and Behavioral Statistics, 40 (3), 307-340. \doi{10.3102/1076998615583902}
}
\author{
Xu Qin and Guanglei Hong
}
